+++ exam-prep-py/get_questions.py
@@ -1,11 +1,14 @@
-from flask import Flask, jsonify
+import os
+from flask import Flask, Blueprint, jsonify
 from flask_cors import CORS
 import json
 
 app = Flask(__name__)
 CORS(app)  # This line enables CORS globally for all domains, fine for development but consider adjusting for production.
 
-@app.route('/get-questions-<module>-<examType>', methods=['GET'])
+questions_bp = Blueprint('questions', __name__)
+
+@questions_bp.route('/get-questions-<module>-<examType>', methods=['GET'])
 def get_questions(module, examType):
     try:
         # Use f-string to dynamically construct the file path
@@ -17,4 +20,4 @@ def get_questions(module, examType):
         return jsonify({"error": f"Questions file for {module} and {examType} not found."}), 404
 
 if __name__ == '__main__':
-    app.run(debug=True, port=5006)
+    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 5000)))
