--- exam-prep-py/howto.md
@@ -1,59 +1,0 @@
-Reducing calls to the OpenAI API while still generating diverse questions for students is a thoughtful approach to managing API usage costs and efficiency. Generating a larger batch of questions in advance and then randomly selecting a subset for each student's use can indeed be a smart strategy. Here are several reasons why this approach makes sense and how you can implement it:
-
-### Benefits:
-
-1. **Reduced API Calls:** By generating a large batch of questions in advance, you minimize the frequency of API calls, which can help manage costs and avoid hitting rate limits.
-2. **Improved Performance:** Serving pre-generated questions from a local database or file system is significantly faster than generating them on-demand, leading to a better user experience.
-3. **Diverse Question Sets:** Randomly selecting from a large pool ensures that students are less likely to see the same questions repeatedly, enhancing the learning experience.
-
-### Implementation Steps:
-
-1. **Batch Generation:**
-   - Periodically (e.g., weekly or monthly, depending on your application's usage patterns), generate a large batch of questions using the OpenAI API. 
-   - Aim for a number that comfortably exceeds your expected unique visitor count within the period before the next batch generation. Generating 60-100 questions might be a good start, depending on your specific needs.
-
-2. **Storage:**
-   - Store these questions in a database or a structured file (like JSON). Ensure each question has a unique identifier for easy retrieval.
-
-3. **Random Selection:**
-   - When a student starts a quiz or session, randomly select 20 questions from your pool.
-   - To enhance diversity further, you can tag questions by topic or difficulty and select randomly within those tags to ensure a balanced quiz.
-
-4. **Caching and Updating:**
-   - Cache these questions on the server or client side to reduce load times for subsequent accesses.
-   - Regularly update the question pool to keep the content fresh and engaging.
-
-### Example Code Snippet for Random Selection:
-
-Assuming you have a JSON file with questions, here’s a simple way to randomly select questions in Python:
-
-```python
-import json
-import random
-
-def get_random_questions(filename, num_questions=20):
-    try:
-        with open(filename, 'r', encoding='utf-8') as file:
-            questions = json.load(file)
-        selected_questions = random.sample(questions, num_questions)
-        return selected_questions
-    except FileNotFoundError:
-        print("Question file not found.")
-        return []
-    except ValueError:
-        print("Not enough questions to select from.")
-        return []
-
-# Example usage
-questions = get_random_questions('generated_content.json', 20)
-for question in questions:
-    print(question)
-```
-
-### Considerations:
-
-- **Quality Control:** Initially, review the generated questions for quality and relevance. Over time, you can automate quality checks or rely on user feedback to filter out or flag less useful questions.
-- **User Feedback:** Implement functionality for students to report errors or provide feedback on questions. This helps you refine the question pool over time.
-- **Ethics and Fairness:** Ensure that the questions generated and the manner in which they are selected do not introduce biases or unfair advantages.
-
-By following this approach, you can efficiently utilize the OpenAI API to create a dynamic and engaging learning experience for students while managing API usage and costs effectively.
