--- exam-prep-py/notes.txt
@@ -1,5 +1,0 @@
-curl -X POST http://127.0.0.1:5003/generate-question \
-     -H "Content-Type: application/json" \
-     -d "{\"topic\": \"Cryptography\"}"
-
-
