+++ exam-prep-py/requirements.txt
@@ -1,0 +1,4 @@
+Flask==3.0.2
+Flask_Cors==4.0.0
+openai==1.16.2
+gunicorn==20.1.0
