+++ exam-prep-py/runtime.txt
@@ -1,0 +1,1 @@
+python-3.10.14
